// src/api/node/types.ts
/* eslint-disable */
// tslint:disable
/**
 * Yeying API
 * 
 *
 * OpenAPI spec version: 1.0
 * Contact: yeying.community@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator+.
 * https://github.com/karlvr/openapi-generator-plus
 * Do not edit the class manually.
 */

import { Express } from 'express'
import { Api } from '../../models'

export interface NodeApi {
	nodeHealthCheck: (request: Api.NodeHealthCheckRequest) => Promise<NodeHealthCheckResponse>
	nodeWhoami: (request: Api.NodeWhoamiRequest) => Promise<NodeWhoamiResponse>
}

export type NodeHealthCheckResponse = NodeHealthCheck200Response | NodeHealthCheckDefaultResponse

export interface NodeHealthCheck200Response {
	status: 200
	body: Api.NodeHealthCheckResponse
	headers?: never
}

export interface NodeHealthCheckDefaultResponse {
	/* Catch-all response */
	status: 'default'
	actualStatus: number
	body: Api.RpcStatus
	headers?: never
}

export type NodeWhoamiResponse = NodeWhoami200Response | NodeWhoamiDefaultResponse

export interface NodeWhoami200Response {
	status: 200
	body: Api.NodeWhoamiResponse
	headers?: never
}

export interface NodeWhoamiDefaultResponse {
	/* Catch-all response */
	status: 'default'
	actualStatus: number
	body: Api.RpcStatus
	headers?: never
}